Refactors routes
